Move CMerkleBlock and CPartialMerkleTree to their own file
