+++ src/test/pmt_tests.cpp
@@ -2,9 +2,11 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "main.h"
+#include "merkleblock.h"
 #include "serialize.h"
+#include "streams.h"
 #include "uint256.h"
+#include "version.h"
 
 #include <vector>
